+++ app/src/main/AndroidManifest.xml
@@ -9,8 +9,6 @@
     <uses-feature android:name="android.hardware.camera" />
     <uses-feature android:name="android.hardware.camera.autofocus" />
 
-
-
     <application
         android:allowBackup="true"
         android:icon="@drawable/ic_launcher"
@@ -19,7 +17,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name="activities.DisclaimerActivity"
+            android:name="com.derma.app.activities.DisclaimerActivity"
             android:label="@string/activity_name_classification"
             android:noHistory="true"
             android:screenOrientation="portrait">
@@ -29,22 +27,22 @@
             </intent-filter>
         </activity>
 
-        <activity android:name="activities.ClassifierActivity">
+        <activity android:name="com.derma.app.activities.ClassifierActivity">
         </activity>
 
-        <activity android:name="activities.AboutActivity">
+        <activity android:name="com.derma.app.activities.AboutActivity">
         </activity>
 
-        <activity android:name="activities.InfoPageActivity">
+        <activity android:name="com.derma.app.activities.InfoPageActivity">
         </activity>
 
-        <activity android:name="activities.MainActivity">
+        <activity android:name="com.derma.app.activities.MainActivity">
         </activity>
 
-        <activity android:name="activities.SharingActivity">
+        <activity android:name="com.derma.app.activities.SharingActivity">
         </activity>
 
-        <activity android:name="activities.ResultActivity">
+        <activity android:name="com.derma.app.activities.ResultActivity">
         </activity>
 
         <!--<activity-->
