+++ app/src/main/assets/nv_labels.txt
@@ -1,2 +1,2 @@
-nv
 non-nv
+nv
