--- app/src/main/java/activities/AboutActivity.java
@@ -1,42 +1,0 @@
-package activities;
-
-import android.Manifest;
-import android.app.Activity;
-
-import android.content.Intent;
-import android.content.pm.PackageManager;
-import android.os.Build;
-import android.os.Bundle;
-import android.view.View;
-
-import com.derma.app.R;
-
-
-
-public class AboutActivity extends Activity {
-    private static final String PERMISSION_CAMERA = Manifest.permission.CAMERA;
-    private static final int PERMISSIONS_REQUEST = 1;
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.about_layout);
-    }
-
-    public void pressOK(View v) {
-        Intent intent = new Intent(this, MainActivity.class);
-        startActivity(intent);
-    }
-
-    public void consentAction(View v) {
-        Intent intent = new Intent(this, MainActivity.class);
-        startActivity(intent);
-    }
-
-    private boolean hasPermission() {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            return checkSelfPermission(PERMISSION_CAMERA) == PackageManager.PERMISSION_GRANTED;
-        } else {
-            return true;
-        }
-    }
-}
