+++ app/src/main/java/com/derma/app/activities/InfoPageActivity.java
@@ -1,0 +1,54 @@
+package com.derma.app.activities;
+import android.Manifest;
+import android.app.Activity;
+
+import android.content.Intent;
+import android.content.pm.PackageManager;
+import android.os.Build;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Toast;
+import com.derma.app.R;
+
+public class InfoPageActivity extends Activity {
+    private static final String PERMISSION_CAMERA = Manifest.permission.CAMERA;
+    private static final int PERMISSIONS_REQUEST = 1;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.info_page);
+
+    }
+
+    public void pressOK(View v) {
+        Intent intent = new Intent(this, MainActivity.class);
+        startActivity(intent);
+    }
+
+    private boolean hasPermission() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+            return checkSelfPermission(PERMISSION_CAMERA) == PackageManager.PERMISSION_GRANTED;
+        } else {
+            return true;
+        }
+    }
+    public void clickBack(View v) {
+        Intent intent = new Intent(this, MainActivity.class);
+        startActivity(intent);
+    }
+
+    private void requestPermission() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+            if (shouldShowRequestPermissionRationale(PERMISSION_CAMERA)) {
+                Toast.makeText(
+                        InfoPageActivity.this,
+                        "Camera permission is required for this demo",
+                        Toast.LENGTH_LONG)
+                        .show();
+            }
+            requestPermissions(new String[] {PERMISSION_CAMERA}, PERMISSIONS_REQUEST);
+        }
+    }
+
+}
