+++ app/src/main/java/com/derma/app/classifier/Stage.java
@@ -1,0 +1,8 @@
+package com.derma.app.classifier;
+
+public enum Stage {
+    ORIGINAL,
+    CROPPED,
+    NV_ClASSIFIED,
+    CLASSIFIED
+}
