--- app/src/main/java/org/tensorflow/lite/examples/classification/Stage.java
@@ -1,8 +1,0 @@
-package org.tensorflow.lite.examples.classification;
-
-public enum Stage {
-    ORIGINAL,
-    CROPPED,
-    NV_ClASSIFIED,
-    CLASSIFIED
-}
