--- app/src/main/java/org/tensorflow/lite/examples/classification/tflite/ClassifierNVDetNet.java
@@ -1,99 +1,0 @@
-/* Copyright 2019 The TensorFlow Authors. All Rights Reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-==============================================================================*/
-
-package org.tensorflow.lite.examples.classification.tflite;
-
-import android.app.Activity;
-import java.io.IOException;
-
-/** This TensorFlowLite classifier works with the float MobileNet model. */
-public class ClassifierNVDetNet extends Classifier {
-
-    /** MobileNet requires additional normalization of the used input. */
-    private static final float IMAGE_MEAN = 159.88f;
-    private static final float IMAGE_STD = 46.75f;
-
-    /**
-     * An array to hold inference results, to be feed into Tensorflow Lite as outputs. This isn't part
-     * of the super class, because we need a primitive array here.
-     */
-    private float[][] labelProbArray = null;
-
-    /**
-     * Initializes a {@code ClassifierSkinDetNet}.
-     *
-     * @param activity
-     */
-    public ClassifierNVDetNet(Activity activity, Device device, int numThreads)
-            throws IOException {
-        super(activity, device, numThreads);
-        labelProbArray = new float[1][getNumLabels()];
-    }
-
-    @Override
-    public int getImageSizeX() {
-        return 75;
-    }
-
-    @Override
-    public int getImageSizeY() {
-        return 100;
-    }
-
-    @Override
-    protected String getModelPath() {
-        // you can download this file from
-        // see build.gradle for where to obtain this file. It should be auto
-        // downloaded into assets.
-        return "model_nv.tflite";
-    }
-
-    @Override
-    protected String getLabelPath() {
-        return "nv_labels.txt";
-    }
-
-    @Override
-    protected int getNumBytesPerChannel() {
-        return 4; // Float.SIZE / Byte.SIZE;
-    }
-
-    @Override
-    protected void addPixelValue(int pixelValue) {
-        imgData.putFloat((((pixelValue >> 16) & 0xFF) - IMAGE_MEAN) / IMAGE_STD);
-        imgData.putFloat((((pixelValue >> 8) & 0xFF) - IMAGE_MEAN) / IMAGE_STD);
-        imgData.putFloat(((pixelValue & 0xFF) - IMAGE_MEAN) / IMAGE_STD);
-    }
-
-    @Override
-    protected float getProbability(int labelIndex) {
-        return labelProbArray[0][labelIndex];
-    }
-
-    @Override
-    protected void setProbability(int labelIndex, Number value) {
-        labelProbArray[0][labelIndex] = value.floatValue();
-    }
-
-    @Override
-    protected float getNormalizedProbability(int labelIndex) {
-        return labelProbArray[0][labelIndex];
-    }
-
-    @Override
-    protected void runInference() {
-        tflite.run(imgData, labelProbArray);
-    }
-}
