+++ app/src/main/res/layout/activity_camera.xml
@@ -32,7 +32,7 @@
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:background="@android:color/black"
-            tools:context="org.tensorflow.lite.examples.classification.activities.CameraActivity" />
+            tools:context="org.tensorflow.lite.examples.classifier.activities.CameraActivity" />
 
 
         <androidx.appcompat.widget.Toolbar
